+++ pkg/util/convert.go
@@ -10,3 +10,13 @@ func MapToSlice[V comparable](m map[V]struct{}) []V {
 
 	return res
 }
+
+// SliceToMap converts any slice with comparable elements to a map.
+func SliceToMap[V comparable](s []V) map[V]struct{} {
+	res := make(map[V]struct{}, len(s))
+	for _, v := range s {
+		res[v] = struct{}{}
+	}
+
+	return res
+}